feat: setup locale at app startup by locatoin path. add base Url to links
load products menu with current locale
