+++ client-app/core/api/graphql/common/queries/getMenus/index.ts
@@ -1,6 +1,6 @@
 import client from "@core/api/graphql/graphql-client";
 import { MenuLinkListType, Query, QueryMenusArgs } from "@core/api/graphql/types";
-import { locale, storeId } from "@core/constants";
+import { storeId } from "@core/constants";
 import getMenusQueryDocument from "./getMenus.graphql";
 
 export default async function getMenus(payload?: Omit<QueryMenusArgs, "storeId">): Promise<MenuLinkListType[]> {
@@ -8,7 +8,6 @@ export default async function getMenus(payload?: Omit<QueryMenusArgs, "storeId">
     query: getMenusQueryDocument,
     variables: {
       storeId,
-      cultureName: locale,
       ...payload,
     },
   });
