+++ client-app/core/utilities/menu/index.ts
@@ -3,7 +3,7 @@ import { MenuLinkType } from "@/core/api/graphql/types";
 
 export const menu: Record<string, MenuLinkType[]> = {};
 
-export async function initMenu(): Promise<Record<string, MenuLinkType[]>> {
-  (await getMenus()).forEach((item) => (menu[`${item.name}`] = item.items || []));
+export async function initMenu(cultureName: string): Promise<Record<string, MenuLinkType[]>> {
+  (await getMenus(cultureName)).forEach((item) => (menu[`${item.name}`] = item.items || []));
   return menu;
 }
