+++ client-app/router.ts
@@ -1,13 +1,6 @@
 import { createRouter, createWebHistory, RouteRecordRaw } from "vue-router";
 import SeoUrl from "@core/seo-routes.enum";
-import {
-  i18n,
-  defaultLocale,
-  supportedLocales,
-  setI18nLocale,
-  loadLocaleMessages,
-  addLocaleAliasToRoutes,
-} from "./i18n";
+import { addLocaleAliasToRoutes } from "./i18n";
 
 // Pages
 const Home = () => import("./pages/home/home.vue");
@@ -82,7 +75,7 @@ if (import.meta.env.MODE === "development") {
 }
 
 // todo: move to plugin
-export default function setupRouter() {
+export default function setupRouter(baseUrl?: string) {
   addLocaleAliasToRoutes(routes);
 
   // Router definition
@@ -90,7 +83,7 @@ export default function setupRouter() {
     routes,
 
     // History mode
-    history: createWebHistory(),
+    history: createWebHistory(baseUrl),
 
     // Setup scroll behavior on route change
     scrollBehavior(to, from, savedPosition) {
@@ -103,36 +96,6 @@ export default function setupRouter() {
         };
       }
     },
-  });
-
-  router.beforeEach(async (to, from, next) => {
-    let locale = to.path.split("/")[1];
-
-    // use saved or default locale if path locale is not exists in available
-    if (!supportedLocales.includes(locale)) {
-      const savedLocale = localStorage.getItem("locale") as string;
-
-      if (supportedLocales.includes(savedLocale)) {
-        locale = savedLocale;
-      } else {
-        locale = defaultLocale;
-      }
-
-      if (locale !== defaultLocale) {
-        const newPath = `/${locale}${to.path.charAt(0) != "/" ? "/" : ""}${to.path}`;
-        return next({ path: newPath, query: to.query, hash: to.hash });
-      }
-    }
-
-    // load locale messages if not yet
-    if (i18n && !i18n.global.availableLocales.includes(locale)) {
-      await loadLocaleMessages(locale);
-    }
-
-    // set i18n language
-    setI18nLocale(locale);
-
-    return next();
   });
 
   return router;
